requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openslide-python==1.3.1
Pillow>=10.0.0
google-cloud-storage>=2.10.0
